% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{conditionValue}
\alias{conditionValue}
\title{Condition value for a sparse matrix}
\usage{
conditionValue(X)
}
\arguments{
\item{X}{A sparse matrix (Matrix).}
}
\value{
A positive numeric value
}
\description{
Computes the condition number for the cross-product of a sparse matrix to itself
}
\details{
The condition number for X'X reflects how close this matrix is to being singular,
i.e. impossible to invert. The larger condition value, the closer it is to being singular.
}
\examples{

}
\seealso{
\code{\link{genomeClustering}}.
}
\author{
Lars Snipen.
}
