% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{corrDist}
\alias{corrDist}
\title{Correlation distances in a sparse matrix}
\usage{
corrDist(X)
}
\arguments{
\item{X}{A sparse matrix (Matrix).}
}
\value{
An nxn distance matrix, when the input sparse matrix has n columns
}
\description{
Computes the correlation distances between pairs of columns of a sparse matrix.
}
\details{
The correlation distance between two columns of a matrix is simply 1 minus the
correlation between them.
}
\seealso{
\code{\link{genomeClustering}}.
}
\author{
Lars Snipen.
}
