% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{getRMSfragments}
\alias{getRMSfragments}
\title{Retrieving RMS fragments from genomes}
\usage{
getRMSfragments(genome, genome.ID = NULL, min.length = 30,
  max.length = 500, left = "G|AATTC", right = "T|TAA",
  verbose = TRUE)
}
\arguments{
\item{genome}{A \code{\link{Fasta}} object with genome data.}

\item{genome.ID}{Unique identifier for each genome, will be added to FASTA-headers (text).}

\item{min.length}{Minimum fragment length (integer).}

\item{max.length}{Maximum fragment length (integer).}

\item{left}{Text with first, long, restriction enzyme cut motif (text).}

\item{right}{Text with second, short, restriction enzyme cut motif (text).}

\item{verbose}{Turn on/off output text during processing (logical).}
}
\value{
A \code{\link{Fasta}} object with all fragment sequences (5'-3').
}
\description{
Retrieves a set of fragments from a genome, given restriction enzyme cutting motifs.
}
\details{
This function is used to find and retrieve all RMS fragments from a genome.
A \code{\link{Fasta}}-object with the genome sequence(s) is required. A \code{genome.ID}, if
supplied, will be added to the FASTA headers of the output, which is useful if you want to trace
the origin of the fragments. All retrieved fragments will then get a FASTA-header starting with the
token <genome.ID>_RMSx, where x is an integer (1,2,...,). This first token is followed by a blank.
This ensures that all first tokens are unique and that the genome of its origin is indicated.

The default restriction enzymes are EcoRI and MseI, with cutting motifs \code{"G|AATTC"} and
\code{"T|TAA"}, respectively. The vertical bar indicates where in the motif the enzyme cuts. The
forward primers are ligated to the left
}
\examples{
# A small genome in this package
xpth <- file.path(path.package("microrms"),"extdata")
genome.file <- file.path(xpth,"GCF_000009605.1_ASM960v1_genomic.fna")

# Read genome, find fragments
gnm <- readFasta(genome.file)
frg <- getRMSfragments(gnm, genome.ID = "my.ID")

# Write to file with writeFasta(frg, out.file = <filename>)

}
\seealso{
\code{\link{RMSdbase}}.
}
\author{
Lars Snipen.
}
