% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normLength}
\alias{normLength}
\title{Length-normalization}
\usage{
normLength(Y, L, max.train = 1000, trunc = 1, verbose = TRUE)
}
\arguments{
\item{Y}{Matrix of readcounts, one row for each fragment cluster, one column for each sample.}

\item{L}{Fragment lengths.}

\item{max.train}{Integer indicating maximum number of data used to train the loess model}

\item{trunc}{Lowest readcount considered as valid when training loess model}

\item{verbose}{Logical, if TRUE text is written to the Console during computations.}
}
\value{
A matrix same size as \code{Y} with corrected readcounts.
}
\description{
Reducing the RMS readcount bias that is due to fragment length.
}
\details{

}
\examples{
more here.

}
\seealso{
more here.
}
\author{
Lars Snipen.
}
