% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruneSamples}
\alias{pruneSamples}
\title{Discarding samples}
\usage{
pruneSamples(rms.obj, mapping.threshold = 0, verbose = TRUE)
}
\arguments{
\item{rms.obj}{A \code{list} with the tables \code{Sample.tbl}, see details below.}

\item{mapping.threshold}{Minimum fraction of reads being mapped.}

\item{verbose}{Turn on/off output text during processing (logical).}
}
\value{
An RMS object where the \code{Sample.tbl} and \code{Readcount.mat} have been pruned, i.e.
some samples have been discarded.
}
\description{
Discarding samples having too few reads mapped.
}
\details{
The \code{rms.obj} must be a list with the required data structures. It must have a 
\code{Sample.tbl}, which is added to such an object with \code{\link{addSampleTable}}. It must
also have a \code{Readcount.mat}, which is added by the \code{\link{readMapper}}.

The pruning means sample where the fraction of mapped reads to the total number of reads is below
the threshold set by \code{mapping.threshold}. If very few reads are mapped, it probably indicates
none of the genomes are present in the sample, and even if abundance estimates can be made they are
also very unreliable in such cases.
}
\examples{

}
\seealso{
\code{\link{RMSobject}}, \code{\link{readMapper}}, \code{\link{rmscols}}.
}
\author{
Lars Snipen.
}
