% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{readMapper}
\alias{readMapper}
\title{Mapping reads}
\usage{
readMapper(sample.files, centroids.file, identity, tmp.dir = ".",
  threads = 1, min.length = 30, unmapped.dir = NULL,
  verbose = TRUE)
}
\arguments{
\item{sample.files}{A vector of filenames, see details below.}

\item{centroids.file}{A filename, see details below.}

\item{identity}{Identity (numeric, 0.0-1.0) threshold for mapping.}

\item{threads}{Number of threads to be used by VSEARCH.}

\item{min.length}{Minimum fragment length (bases).}

\item{unmapped.dir}{If supplied, a directory to put non-mapped reads.}

\item{verbose}{Logical, if TRUE text is written to the Console during computations.}
}
\value{
A
}
\description{
Mapping reads to an RMS database using VSEARCH
}
\details{
The \code{sample.files} must be a vector of (full path) names of FASTA-files containing
the reads from the samples. The \code{centroids.file} is the name of a FASTA-file with the
database centroids to map against. Thus, the reads from each sample will be mapped to these
sequences, using the \code{identity} threshold. This results in a table of read-counts. The \code{sample.files}
are used as column names in this table. The first token in the Headers of the \code{centroids.file} are
used as row names.
}
\examples{
more here.

}
\seealso{
This is version latest
}
\author{
Lars Snipen.
}
