% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{rmscols}
\alias{rmscols}
\title{Estimating microbiota composition}
\usage{
rmscols(Y, X, trim = 0, reltol = 1e-06, verbose = TRUE)
}
\arguments{
\item{Y}{A matrix of read counts, a column for each sample.}

\item{X}{A matrix of amplicon cluster copy numbers, a column for each genome (cluster).}

\item{trim}{Fraction of extreme readcounts to discard when fitting linear model.}

\item{verbose}{Logical, if TRUE text is written to the Console during computations.}
}
\value{
A matrix of the fraction for each genome.
}
\description{
Estimates the fraction of each genome in a sample, based on read counts and copy
numbers for each amplicon cluster.
}
\details{
more here.
}
\examples{
more here.

}
\seealso{
more here.
}
\author{
Lars Snipen.
}
