% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{addSampleTable}
\alias{addSampleTable}
\title{Adds sample table}
\usage{
addSampleTable(rms.obj, sample.tbl)
}
\arguments{
\item{rms.obj}{An \code{list} with RMS data structures, see \code{\link{RMSobject}}.}

\item{sample.tbl}{A table with sample metadata (data.frame or tibble).}
}
\value{
A \code{list} similar to the input \code{rms.obj}, but with one new element added.
}
\description{
Adds a sample table to an rms object
}
\details{
This small function just adds a table with metadata about samples to an 
already existing \code{rms.obj}. The latter is a \code{list}, and this function only
ensures the added element is correctly named \code{Sample.tbl}.

If the supplied \code{rms.obj} already contains a \code{Sample.tbl}, it is simply replaced by the new \code{sample.tbl}.

The \code{sample.tbl} must contain at least the two columns \code{sample_id} and 
\code{fasta_file}. The first is a unique text to identify each sample, the latter is
the name of the fasta file with processed reads, see \code{\link{readMapper}}.
}
\examples{

}
\seealso{
\code{\link{RMSobject}}.
}
\author{
Lars Snipen.
}
