% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{getRMSfragments}
\alias{getRMSfragments}
\title{Retrieving RMS fragments from genomes}
\usage{
getRMSfragments(genome, genome.ID = NULL, min.length = 30,
  max.length = 500, left = "GAATTC", right = "TTAA",
  verbose = TRUE, trim = TRUE)
}
\arguments{
\item{genome}{A \code{\link{Fasta}} object with genome data.}

\item{genome.ID}{Unique text identifier for each genome, will be added to FASTA-headers.}

\item{min.length}{Minimum amplicon length (bases).}

\item{max.length}{Maximum amplicon length (bases).}

\item{left}{Text with first, long, restriction enzyme cut motif. Deafult is the EcoRI.}

\item{right}{Text with second, short, restriction enzyme cut motif. Deafult is the MseI.}

\item{verbose}{Logical to turn on/off output text during processing.}

\item{trim}{Logical indicating if the restriction motifs above should be trimmed off the ends of the fragments. Default is \code{TRUE}}
}
\value{
A \code{\link{Fasta}} object with all fragment sequences (5'-3').
}
\description{
Retrieves a set of fragments from a genome, given restriction enzyme cutting motifs.
}
\details{
This function is used to find and retrieve all RMS fragments from a genome. A \code{\link{Fasta}}-object
with the genome sequence(s) is required. A \code{genome.ID} may/should be supplied, and will be added to the header-lines of the output
\code{\link{Fasta}} object to identify the origin of the fragments. All retrieved fragments will then get
a FASTA-header starting with the token <genome.ID>_RMSx, where x is an integer (1,2,...,). This first token is followed
by a blank. This ensures that all first tokens are unique and that the genome of its origin is indicated.

The default restriction enzymes are EcoRI and MseI,
with cutting motifs \code{"GAATTC"} and \code{"TTAA"}, respectively. Change cutting motifs accordingly if
you use other restriction enzymes.
}
\examples{
more here.

}
\seealso{
more here.
}
\author{
Lars Snipen.
}
