% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{getRMSfragments}
\alias{getRMSfragments}
\title{Retrieving RMS fragments from genomes}
\usage{
getRMSfragments(
  genome,
  genome.id,
  min.length = 30,
  max.length = 500,
  left = "G|AATTC",
  right = "T|TAA",
  verbose = TRUE
)
}
\arguments{
\item{genome}{A table (fasta object) with genome data.}

\item{genome.id}{Unique identifier for each genome, will be added to FASTA-headers (text).}

\item{min.length}{Minimum fragment length (integer).}

\item{max.length}{Maximum fragment length (integer).}

\item{left}{Text with first, long, restriction enzyme cut motif (text).}

\item{right}{Text with second, short, restriction enzyme cut motif (text).}

\item{verbose}{Turn on/off output text during processing (logical).}
}
\value{
A \code{\link{tibble}} with with all fragment sequences (5'-3') in FASTA format.
}
\description{
Retrieves a set of fragments from a genome, given restriction enzyme cutting motifs.
}
\details{
This function is used to find and retrieve all RMS fragments from a \code{genome}.
This is a \code{\link{tibble}} with sequence data in FASTA format, see \code{\link{readFasta}}.
In addition, a \code{genome.id} is required,
which is a text unique to each genome to be analyzed. This \code{genome.id} will be added to the
fasta headers of the output, and all headers start with the token <genome.id>_RMSx, where x is
an integer (1,2,...,). This first token is followed by a blank. This ensures that all first tokens
are unique and that the genome of its origin is indicated for all fragments

The default restriction enzymes are EcoRI and MseI, with cutting motifs \code{"G|AATTC"} and
\code{"T|TAA"}, respectively. The vertical bar indicates the cut site in the motif.
}
\examples{
# A small genome in this package
xpth <- file.path(path.package("microrms"),"extdata")
genome.file <- file.path(xpth,"GCF_000009605.1_ASM960v1_genomic.fna")

# Read genome, find fragments
gnm <- readFasta(genome.file)
frg <- getRMSfragments(gnm, "genome_1")

# Write to file with writeFasta(frg, out.file = <filename>)

}
\seealso{
\code{\link{RMSobject}}.
}
\author{
Lars Snipen.
}
