% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normLength}
\alias{normLength}
\title{Length-normalization}
\usage{
normLength(rms.obj, max.train = 1000, readcount.low = 1,
  verbose = TRUE)
}
\arguments{
\item{rms.obj}{A \code{list} with RMS data structures, see \code{\link{RMSobject}}.}

\item{max.train}{Integer indicating maximum number of data points used to train the loess model.}

\item{readcount.low}{Lowest readcount considered as valid when training loess model.}

\item{verbose}{Logical, if TRUE text is written to the Console during computations.}
}
\value{
An rms object similar to \code{rms.obj}, but where the \code{Readcount.mat}
matrix has been normalized, columnwise.
}
\description{
Reducing the RMS readcount bias that is due to fragment length.
}
\details{
RMS amplicon signals (readcounts) usually have some bias due to varying lengths. This
function attempts to remove this bias. It has been described in detail in Snipen et al, 2019.
}
\examples{
See tutorial.

}
\author{
Lars Snipen.
}
