% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{readMapper}
\alias{readMapper}
\title{Mapping reads}
\usage{
readMapper(sample.files, centroids.file, identity = 0.99, threads = 1,
  min.length = 30, verbose = TRUE, tmp.dir = NULL,
  unmapped.dir = NULL)
}
\arguments{
\item{sample.files}{A vector of FASTA-filenames (text).}

\item{centroids.file}{A FASTA-filename (text).}

\item{identity}{Identity threshold for mapping (0.0-1.0).}

\item{threads}{Number of threads to be used by vsearch (integer).}

\item{min.length}{Minimum fragment length used by vsearch (integer).}

\item{verbose}{Turn on/off output text during processing (logical).}

\item{tmp.dir}{If supplied, the output directory used by vsearch.}

\item{unmapped.dir}{If supplied, a directory to output unmapped reads.}
}
\value{
A matrix of read-counts, with one row for each fragment cluster and one column for each
sample file.
}
\description{
Mapping reads to an RMS database using VSEARCH
}
\details{
The \code{sample.files} must be a vector of names of FASTA-files containing
the reads from the samples. The \code{centroids.file} is the name of a FASTA-file with the
database centroids to map against. Thus, the reads from each sample will be mapped to these
sequences, using the \code{identity} threshold.

This results in a matrix of read-counts. The column names of this matrix are the filenames in
\code{sample.files}, after the removal of the path and the file extension. The first token in
the Headers of the \code{centroids.file} are used as row names.
}
\examples{
\dontrun{
xpth <- file.path(path.package("microrms"),"extdata")
smp.files <- file.path(xpth, list.files(xpth, pattern = "Sample"))
ctr.file <- file.path(xpth, "centroids.fasta")

read.counts <- readMapper(smp.files, ctr.file)
}

}
\seealso{
\code{\link{RMSdbase}}, \code{\link{rmscols}}.
}
\author{
Lars Snipen.
}
