% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{readMapper}
\alias{readMapper}
\title{Mapping reads}
\usage{
readMapper(
  rms.obj,
  fa.dir,
  vsearch.exe = "vsearch",
  identity = 0.99,
  threads = 1,
  min.length = 30,
  verbose = TRUE
)
}
\arguments{
\item{rms.obj}{A \code{list} with the tables \code{Sample.tbl} and \code{Cluster.tbl}, see details below.}

\item{fa.dir}{A path to where the fasta files with reads are located.}

\item{vsearch.exe}{Text with the VSEARCH executable command.}

\item{identity}{Identity threshold for mapping (0.0-1.0).}

\item{threads}{Number of threads to be used by vsearch (integer).}

\item{min.length}{Minimum fragment length used by vsearch (integer).}

\item{verbose}{Turn on/off output text during processing (logical).}
}
\value{
An RMS object with the matrix \code{Readcount.mat} added. Also, two new columns, 
\code{reads_total} and \code{reads_mapped}, have been added to the \code{Sample.tbl}.
}
\description{
Mapping reads to clusters in an RMS object using VSEARCH.
}
\details{
The \code{rms.obj} must be a list with the required data structures for mapping
reads, i.e. it must contain a \code{Sample.tbl}, see \code{\link{addSampleTable}}. Note
that this table must have a column named \code{fasta_file} naming the fasta files with
the reads to be mapped, one file for each sample. The argument \code{fa.dir} is used to
specify the path to these files.

The \code{rms.obj} must also contain a \code{Cluster.tbl}, see \code{\link{RMSobject}}.
The reads from each sample will be mapped to the cluster
sequences, using the \code{identity} threshold.

This results in a matrix of read counts, with one row for each fragment cluster and one column
for each sample. The column names of this matrix are the \code{sample_id}
texts in the \code{Sample.tbl}. The first token in the \code{Header} of the \code{Cluster.tbl} 
are used as row names.

The \code{vsearch.exe} is the exact command to invoke the VSEARCH software. This is normally just "vsearch", 
but if you run this as a singularity container (or any other container) it may be something like
"singularity exec <container_name> vsearch".
}
\examples{

}
\seealso{
\code{\link{RMSobject}}, \code{\link{rmscols}}.
}
\author{
Lars Snipen.
}
