% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{readMapper}
\alias{readMapper}
\title{Mapping reads}
\usage{
readMapper(rms.obj, fa.dir, identity = 0.99, threads = 1,
  min.length = 30, verbose = TRUE)
}
\arguments{
\item{rms.obj}{A \code{list} with the tables \code{Sample.tbl} and \code{Cluster.tbl}, see details below.}

\item{fa.dir}{A path to where the fasta files with reads are located.}

\item{identity}{Identity threshold for mapping (0.0-1.0).}

\item{threads}{Number of threads to be used by vsearch (integer).}

\item{min.length}{Minimum fragment length used by vsearch (integer).}

\item{verbose}{Turn on/off output text during processing (logical).}
}
\value{
A matrix of read-counts, with one row for each fragment cluster and one column for each
sample file.
}
\description{
Mapping reads to clusters in an RMS object using VSEARCH
}
\details{
The \code{sample.files} must be a vector of names of FASTA-files containing
the reads from the samples. The \code{centroids.file} is the name of a FASTA-file with the
database centroids to map against. Thus, the reads from each sample will be mapped to these
sequences, using the \code{identity} threshold.

This results in a matrix of read-counts. The column names of this matrix are the filenames in
\code{sample.files}, after the removal of the path and the file extension. The first token in
the Headers of the \code{centroids.file} are used as row names.
}
\examples{
See tutorial.

}
\seealso{
\code{\link{RMSobject}}, \code{\link{rmscols}}.
}
\author{
Lars Snipen.
}
