% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbase.R
\name{RMSdbase}
\alias{RMSdbase}
\title{Constructing RMS database}
\usage{
RMSdbase(fragment.file, identity = 0.97, min.length = 30,
  max.length = 500, verbose = TRUE, threads = 1)
}
\arguments{
\item{fragment.file}{A FASTA-file with all the RMS fragments.}

\item{identity}{The sequence identity for clustering fragments.}

\item{min.length}{Minimum fragment length (bases).}

\item{max.length}{Maximum fragment length (bases).}

\item{verbose}{Logical to turn on/off output text during processing.}

\item{threads}{Number of threads to be used by \code{vsearch}.}
}
\value{
A list with the following objects: \code{Cluster.tbl}, \code{Cpn.mat}
and \code{Genome.tbl}.

The \code{Cluster.tbl} is a \code{\link{tibble}} with data about all fragment clusters.
It contains columns with data about each cluster, including the centroid \code{Sequence}
and its \code{Header}, making it possible to write the table to a FASTA-file using
\code{\link{writeFasta}}.

The \code{Cpn.mat} is the copy number matrix, implemented as a sparse dgeMatrix from the
\code{\link{Matrix}} package. It has one row for each fragment cluster and one column
for each genome. This is the central data structure for de-convolving the genome
content from read-count data, see \code{\link{deconvolve}}.

The \code{Genome.tbl} is a \code{\link{tibble}} with a row for each genome, and the
total number and number of unique fragment clusters for each genome.
}
\description{
Constructs a database of information about a set of RMS fragments.
}
\details{
The \code{fragment.file} must be a FASTA-file containing all fragments to be clustered. The first token
of each Header-line must indicate the genome of its origin. It must follow the pattern <genome.ID>_RMSx, where
<genome.ID> is some text unique to each genome, and x
is some integer. See \code{\link{getRMSfragments}} for how to make such FASTA-files.
}
\examples{
more here.

}
\seealso{
This is version latest
}
\author{
Lars Snipen.
}
